backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
yt-dlp>=2023.7.6
imageio-ffmpeg>=0.5.1
